Standalone

+++ src/app/consumers/list/list.ts
@@ -1,20 +1,26 @@
-import { Component, inject } from '@angular/core';
+import { Component, inject, OnInit, signal } from '@angular/core';
 import { ConsumerData } from '../model/consumer';
 import { Consumer } from '../consumer';
-import { Router } from '@angular/router';
+import { Router, RouterLink } from '@angular/router';
 import { Unsubscribe } from '../../common/unsubscribe';
+import { MatFormField, MatLabel, MatInput } from '@angular/material/input';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
+import { MatIconButton } from '@angular/material/button';
+import { MatIcon } from '@angular/material/icon';
+import { MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow } from '@angular/material/table';
+import { PhonePipe } from '../../common/phone-pipe';
 
 @Component({
-  selector: 'crm-list',
-  standalone: false,
-  templateUrl: './list.html',
-  styleUrl: './list.scss'
+    selector: 'crm-list',
+    templateUrl: './list.html',
+    styleUrl: './list.scss',
+    imports: [MatFormField, MatLabel, MatInput, ReactiveFormsModule, FormsModule, MatIconButton, RouterLink, MatIcon, MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow, PhonePipe]
 })
-export class List extends Unsubscribe {
+export class List extends Unsubscribe implements OnInit {
   private readonly consumerService = inject(Consumer);
   private readonly router = inject(Router);
 
-  consumers: ConsumerData[] = [];
+  consumers = signal<ConsumerData[]>([]);
   search = '';
   displayedColumns: string[] = ['civility', 'firstname', 'lastname', 'email', 'phone', 'actions'];
 
@@ -28,7 +34,7 @@ export class List extends Unsubscribe {
     }
     this.subscriptionList.push(this.consumerService.find(this.search).subscribe(
       (consumers) => {
-        this.consumers = consumers;
+        this.consumers.set(consumers);
       }
     ));
   }
@@ -38,7 +44,9 @@ export class List extends Unsubscribe {
   }
 
   delete(consumer: ConsumerData): void {
-    this.subscriptionList.push(this.consumerService.remove(consumer.id!).subscribe(
-      () => this.find()));
+    if (confirm(`Are you sure you want to delete ${consumer.firstname} ${consumer.lastname}?`)) {
+      this.subscriptionList.push(this.consumerService.remove(consumer.id!).subscribe(
+        () => this.find()));
+    }
   }
 }
